requirements.txt
djangorestframework
psycopg
pillow